/*/*   This version passed all test cases.   It improves the compareTo method without using the build-in compareTo
Sort an array of numeric strings where each string is a posiive number with anywhere from 1 to 1000000 digits.
Sort the array's elments in non-decreaseing or ascending order of their integer values and return the sorted array.
*****
My idea is to implement my own MyString class which implements Comparable interface.
In the compareTo method of the MyString class, I check if the two strings are the same length.
If they are I compare them at each Char begining from the Left Most one.

It fails one of the test case with 200000 elements due to time out.  got 16.80/20
 */